% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCryptoSymbols.R
\name{getCryptoSymbols}
\alias{getCryptoSymbols}
\title{Recupera a lista de Cryptoativos com a relação USD avaliada no site Yahoo Finance}
\usage{
getCryptoSymbols()
}
\value{
dataframe com as informações dos símbolos dos criptoativos e seu significado
}
\description{
Recupera a lista de Cryptoativos com a relação USD avaliada no site Yahoo Finance
}
\examples{
getCryptoSymbols()
}
